refactor: threading logic of breeze-ui and js
The concept is that any js callback should be run in js thread, and any dom operation should be locked with rt_lock. We post the task to js thread in detached mode, and add the lock to dom operation apis, then it would be safe to operate dom anywhere in js as no js code would block breeze-ui update/render anymore.